+++ src/Libraries/SmartStore.Core/Domain/DataExchange/ExportEnums.cs
@@ -12,7 +12,8 @@ namespace SmartStore.Core.Domain.DataExchange
 		Manufacturer,
 		Customer,
 		Order,
-		NewsLetterSubscription
+		NewsLetterSubscription,
+		ShoppingCartItem
 	}
 
 	/// <summary>